Rom Bank Number changed from byte to short
-In preparation for MBC5 support the romBankNumber is now a short.

Previous MBCs usually used 7 bits. MBC5 uses 9, and therefore a byte is
no longer sufficient.

+++ GameBeak/GameBeak/Memory.h
@@ -23,7 +23,7 @@ class Memory
 	byte beakRom[0x500000]; //64kb for rom
 	string title;
 	bool ramEnabled = false;
-	byte romBankNumber = 0;
+	short romBankNumber = 0;
 	byte ramBankNumber = 0;
 	bool bankingMode = false; //0: Rom 1: Ram
